Migrate user management out of admin namespace
