+++ app/views/admin/index.haml
@@ -1,4 +1,4 @@
 = link_to "View Escalation Requests", admin_escalation_requests_path, class: "btn btn-info"
-= link_to "Manage Users", admin_users_path, class: "btn btn-info"
+= link_to "Manage Users", users_path, class: "btn btn-info"
 = link_to "Update Day Color", edit_day_path, class: "btn btn-info"
 = link_to "Manage Integrations", integrations_path, class: "btn btn-primary"
