+++ app/views/users/_form.haml
@@ -1,4 +1,4 @@
-= form_for [:admin, @user] do |f|
+= form_for @user do |f|
   .form-group
     = f.label :name
     = f.text_field :name, placeholder: "Name", class: "form-control"
