+++ app/views/users/index.haml
@@ -12,7 +12,7 @@
           - @users.each do |user|
             %tr
               %td
-                = link_to user.name, edit_admin_user_path(user)
+                = link_to user.name, edit_user_path(user)
                 &nbsp;
                 - if user.super_admin
                   %i.fa.fa-shield
